+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 - `pagination` builder (https://github.com/edvin/tornadofx/issues/76)
 - Configurable alert dialog 
 - `Node.bindClass` applies an observable style class to a node (https://github.com/edvin/tornadofx/issues/88)
+- Toolbar.spacer and ToolBar.separator builders
 
 ### Changed